+++ simple-stack-example-scoping/src/main/kotlin/com/zhuinden/simplestackexamplescoping/FragmentStateChanger.kt
@@ -38,7 +38,10 @@ class FragmentStateChanger(
                 var fragment: Fragment? = fragmentManager.findFragmentByTag(newKey.fragmentTag)
                 if (newKey == stateChange.topNewState<Any>()) {
                     if (fragment != null) {
-                        if (fragment.isDetached) {
+                        if (fragment.isRemoving) { // Fragments are quirky, they die asynchronously. Ignore if they're still there.
+                            fragment = newKey.newInstance()
+                            replace(containerId, fragment, newKey.fragmentTag)
+                        } else if (fragment.isDetached) {
                             attach(fragment)
                         }
                     } else {