+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 -Simple Stack ?.??.? (????-??-??)
 --------------------------------
 
-- ADD: Safer version of the FragmentStateChanger that handles re-entrancy of `back` and going to the same target as where we were (`simple-stack-example-mvp-fragments`).
+- UPDATE: Safer version of the FragmentStateChanger that handles re-entrancy of `back` and going to the same target as where we were (handle `fragment.isRemoving`).
 
 
 -Simple Stack 1.14.0 (2019-03-16)